Improve VerificationOnboardingStep
